Add more components
