+++ components/ActivityMenu.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import { View, Text } from 'react-native';
+
+export default class ActivityMenu extends Component {
+	render() {
+		return (
+			<View style={{flex: 1, backgroundColor: 'pink'}}>
+				<Text>ActivityMenu Component</Text>
+			</View>
+		);
+	}
+}
