+++ components/ClassroomImage.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import { View, Text } from 'react-native';
+
+export default class ClassroomImage extends Component {
+	render() {
+		return (
+			<View style={{flex: 5, backgroundColor: 'white'}}>
+				<Text>ClassroomImage Component</Text>
+			</View>
+		);
+	}
+}
