+++ components/RightSide.js
@@ -1,33 +1,14 @@
 import React, { Component } from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { View } from 'react-native';
 
-import FeedbackBar from './FeedbackBar';
+import ActivityMenu from './ActivityMenu';
 
 export default function RightSide() {
 	return (
-		<View style={{flex: 5, backgroundColor: 'skyblue'}}>
-			<View style={{flex: 1, flexDirection: 'row'}}>
-				<View style={{flex: 1}}>
-					<Text style={styles.title}>Classroom</Text>
-				</View>
-			</View>
-			<View style={{flex: 8, backgroundColor: 'white'}}>
-				<View style={{flex: 1}}>
-					<FeedbackBar />
-				</View>
-				<View style={{flex: 8}}>
-				</View>
-				<View style={{flex: 3, backgroundColor: 'pink'}}>
-				</View>
+		<View style={{flex: 2, backgroundColor: 'lightblue'}}>
+			<ActivityMenu />
+			<View style={{flex: 9}}>
 			</View>
 		</View>
 	);
 }
-
-const styles = StyleSheet.create({
-	title: {
-		fontWeight: 'bold',
-		fontSize: 50,
-		padding: 20
-	}
-});
